api/comvis/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
numpy==1.24.1
opencv-python==4.6.0.66
pathlib==1.0.1
Pillow==9.3.0
pycparser==2.21
pyheif==0.7.1
requests==2.28.1
urllib3==1.26.13
yarg==0.1.9
